Renamed the common project.
