+++ AudioAnalytics/AudioAnalyzer/AudioAnalyzer.csproj
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Common-Authentication\Common_Authentication.csproj" />
+    <ProjectReference Include="..\..\Common_Utils\Common_Utils.csproj" />
   </ItemGroup>
 
   <ItemGroup>
