+++ AudioAnalytics/AudioAnalyzer/Program.cs
@@ -3,7 +3,7 @@
 
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
-using Common_Authentication;
+using Common_Utils;
 using Microsoft.Azure.EventHubs;
 using Microsoft.Azure.EventHubs.Processor;
 using Microsoft.Azure.Management.Media;
