+++ Live/LiveEventWithDVR/LiveEventWithDVR.csproj
@@ -10,10 +10,10 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="5.0.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
     <PackageReference Include="Microsoft.Identity.Client" Version="4.30.1" />
-  </ItemGroup>
 
+  </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\Common-Authentication\Common_Authentication.csproj" />
+    <ProjectReference Include="..\..\Common_Utils\Common_Utils.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -21,11 +21,11 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-  
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'"> 
-    <RootNamespace>LiveSample</RootNamespace> 
-    <LangVersion>latest</LangVersion> 
-  </PropertyGroup> 
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <RootNamespace>LiveSample</RootNamespace>
+    <LangVersion>latest</LangVersion>
+  </PropertyGroup>
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
