+++ Live/LiveEventWithDVR/Program.cs
@@ -1,4 +1,4 @@
-using Common_Authentication;
+using Common_Utils;
 using Microsoft.Azure.EventHubs;
 using Microsoft.Azure.EventHubs.Processor;
 using Microsoft.Azure.Management.Media;
