+++ Streaming/AssetFilters/AssetFilters.csproj
@@ -7,6 +7,10 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="5.0.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
     <PackageReference Include="Microsoft.Identity.Client" Version="4.30.1" />
+
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common_Utils\Common_Utils.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -26,9 +30,6 @@
     </Content>
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\Common-Authentication\Common_Authentication.csproj" />
-  </ItemGroup>
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
